git-svn-id: http://multitude.googlecode.com/svn/trunk@995 d2709d7f-9d48-0410-97af-374aaeb070bb

+++ Luminous/GLSLProgramObject.hpp
@@ -92,7 +92,7 @@ namespace Luminous
 
     bool loadStrings(const char* vsString, const char* fsString);
 
-    int shaderObjectCount() const { return m_shaderObjects.size(); }
+    int shaderObjectCount() const { return (int) m_shaderObjects.size(); } 
 
     GLuint handle() const { return m_handle; }
